requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
pydantic==2.5.0
python-dotenv==1.0.0
lxml==4.9.3
apscheduler==3.10.4
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
sshtunnel==0.4.0
google-generativeai==0.3.2
